Curtis and Daniel worked on fixing chmod and getting access to work

+++ file.h
@@ -9,6 +9,7 @@ struct file {
 };
 
 
+
 // in-memory copy of an inode
 struct inode {
   uint dev;           // Device number
@@ -20,7 +21,7 @@ struct inode {
   short major;
   short minor;
   short nlink;
-  short permBit;
+  uint permBit;
   uint size;
   uint addrs[NDIRECT+1];
   uint UID;
